+++ application/config/routes.php
@@ -2,3 +2,9 @@
 $route['battle/prebattle'] = 'battle/prebattle';
 
 $route['battle/resolve'] = 'battle/resolve';
+
+$route['battle/start'] = 'battle/start';
+$route['battle/finalize'] = 'battle/finalize';
+$route['battle/report/(:num)'] = 'battle/report/$1';
+$route['war/declare'] = 'war/declare';
+$route['war/scoreboard/(:num)'] = 'war/scoreboard/$1';